[#145] 다이어리 페이지 스켈레톤 적용
