+++ components/skeleton/image/image.module.scss
@@ -1,3 +1,5 @@
+@import '../skeleton-animation.module.scss';
+
 .skeletonContainer {
   display: flex;
   flex-direction: column;
@@ -9,12 +11,12 @@
   height: 64px;
   width: 64px;
   border-radius: 50%;
-  background-color: var(--skeleton-color);
+  @include skeleton-animation;
 }
 
 .skeletonName {
   height: 17px;
   width: 25px;
   border-radius: 10px;
-  background-color: var(--skeleton-color);
+  @include skeleton-animation;
 }
